+++ src/main/java/com/example/startup2/repository/AmountTimeRepository.java
@@ -3,5 +3,5 @@ package com.example.startup2.repository;
 import com.example.startup2.model.AmountTime;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AmountTimeRepository extends CrudRepository<AmountTime, Integer> {
+public interface AmountTimeRepository extends CrudRepository<AmountTime, Long> {
 }
