+++ src/main/java/com/example/startup2/repository/RouteRepository.java
@@ -3,10 +3,9 @@ package com.example.startup2.repository;
 import com.example.startup2.model.Route;
 import org.springframework.data.repository.CrudRepository;
 
-import java.util.List;
 import java.util.Optional;
 
-public interface RouteRepository extends CrudRepository<Route, Integer> {
+public interface RouteRepository extends CrudRepository<Route, Long> {
 
     Optional<Route> findRouteByNumberAndTypeNumber(int number, String typeNumber);
 }
